requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2==2.9.5
python-dotenv==0.21.1
SQLAlchemy==2.0.3
typing_extensions==4.4.0
Werkzeug==2.2.2
